api/requirements.txt
cantools==36.1.0
Django==3.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.4
python-can==3.3.4
